+++ changelog.md
@@ -1,10 +1,14 @@
 ### v2.3.1
 ---
 * **History**
- * Now tracks tier tokens received, group size and instanceMapIDs.
- * Tier tokens received is displayed in the more info details in the voting frame.
- * Removed response text from voting frame more info to make it smaller.
  * The loot history is now enabled by default.
+ * Now tracks tier tokens received, group size and instanceMapIDs.
+ * Tier tokens received from the current instance is displayed under more info in the voting frame.
+ * All tier tokens are displayed in the loot history more info.
+ * *Note: neither of these are backwards compatible with non-english clients, but will show up for all items awarded after this update.*
+ * Total awards are now displayed in the loot history.
+ * Loot History is now sorted by award time by default.
+ * Removed response text from voting frame more info to make it smaller.
 
 
 * **Tier tokens**